refactor(MainTest) : 이름, 횟수, 자동차, 컨트롤러 별 테스트 파일로 나눔
